Refine calculator function after an operation
The calculator can now do an operation, have the outcome applied to the
operand1 variable (as well as in the output text,) and have the user
press on a new digit to do a totally new operation, the calculator
delete's the previous solution (operand1) and the output text is the new
digit.

But, if the user still wishes to use the previous operation's solution
in the next, new operation, they still can!