All solutions are now at most 13 digits long
Using conditions to check if the solution is 13 digits, the maximum
length I want the output to be, along with a loop to delete the last
character of the number until it is 13 digits long, the solution will
now always be 13 or less digits long.

I chose this method over using toFixed() for visual purposes. I also
believe is more consistent!